Move clusterization code to core
